Move shared types in root directory + update imports
