+++ functions/src/api/v1/rooms/service/createTwilioClientToken.ts
@@ -1,5 +1,5 @@
-import { Room } from '../../../../types/Room'
-import { UID } from '../../../../types/uid'
+import { Room } from '../../../../../../types/Room'
+import { UID } from '../../../../../../types/uid'
 import * as AccessToken from 'twilio/lib/jwt/AccessToken'
 import { getTwilioEnv } from '../../../../firebase/env'
