+++ src/services/fetch-token.ts
@@ -1,6 +1,6 @@
 import { getJwtToken } from '../actions/getJwtToken'
-import { UID } from '../../functions/src/types/uid'
-import { JWTToken } from '../../functions/src/types/JWT'
+import { UID } from '../../types/uid'
+import { JWTToken } from '../../types/JWT'
 
 export const fetchToken = async (uid: UID): Promise<JWTToken> => {
     const jwtToken = await getJwtToken(uid)
