+++ types/JoinRoomResponse.ts
@@ -1,0 +1,4 @@
+export interface JoinRoomResponse {
+    jwtAccessToken: string
+    ttl: number
+}
